lib/statement_decoder.py
import re
import json


def decode_line(line, rules):
    for rule in rules:
        pattern_info = json.loads(rule['pattern'])
        pattern = pattern_info['pattern']
        variable_positions = pattern_info.get('variable_positions') or {}

        category = rule['category']

        print(f"pattern: {pattern}")
        print(f"line: {line}")
        result_search = re.search(pattern, line)
        print(f"result_search: {result_search}")

        if not result_search:
            continue

        match = result_search.groups()

        result = {
            'category': category,
            'rule': rule['pattern']
        }

        if category == 'skip':
            return result

        found_all = True

        for variable_name, position in variable_positions.items():
            if position < len(match):
                result[variable_name] = match[position]
            else:
                found_all = False
                break

        if not found_all:
            # need to skip this rule
            continue

        return result

    return None
